Create a Dummy Variable in Seurat Object Metadata with Optional Joining

This function creates a dummy variable in the metadata of a Seurat object. The dummy variable is generated by sampling a specified number of levels and optionally combining it with an existing variable in the metadata.

Usage

make_dummy_variable(
  seu,
  variable_name = "sample",
  number = 2,
  joining_variable = NULL
)

Arguments

seu

A Seurat object. The Seurat object whose metadata will be modified.

variable_name

A character string specifying the name of the dummy variable. Default is "sample". The function will raise an error if a column with this name already exists in the metadata.

number

An integer specifying the range of numbers to sample for the dummy variable. Default is 2.

joining_variable

A character string specifying an existing variable in the metadata to join with the newly created dummy variable. If NULL (the default), no joining is done. If provided and the variable does not exist in the metadata, the function will raise an error.

Value

Returns the modified Seurat object with the new dummy variable in its metadata. If a joining variable is specified, the dummy variable will be concatenated with the values from the joining variable.

Examples

if (FALSE) {
seu <- make_dummy_variable(seu, variable_name = "sample_group", number = 3, joining_variable = "treatment")
}